refactor: 更新首页模块相关图片路径
