+++ src/components/ActivityDetailPopup.vue
@@ -179,7 +179,7 @@ const star = computed(() => props.detail?.score ?? 0)
     height: 44px;
     top: -32px;
     z-index: -1;
-    background: url('/imgs/mask.png');
+    background: url('@/assets/mask.png');
     padding-bottom: 12px;
     .t-icon {
       color: var(--td-font-white-1);
