--- src/mocks/activityMocks.ts
@@ -1,253 +1,0 @@
-export const covers = [
-  '/mock-imgs/activity/cover1.png',
-  '/mock-imgs/activity/cover2.png',
-  '/mock-imgs/activity/cover3.png',
-  '/mock-imgs/activity/cover4.png',
-  'https://tdesign.gtimg.com/mobile/demos/swiper1.png',
-  'https://tdesign.gtimg.com/mobile/demos/swiper2.png',
-  'https://tdesign.gtimg.com/mobile/demos/example1.png',
-  'https://tdesign.gtimg.com/mobile/demos/example2.png',
-  'https://tdesign.gtimg.com/mobile/demos/example3.png',
-]
-
-export const swiperList = [
-  {
-    id: '1',
-    name: '2019 SICC服务设计创新大会',
-    url: '/imgs/activity/sicc2019.png',
-  },
-  {
-    id: '2',
-    name: '2021 SICC服务设计创新大会',
-    url: '/imgs/activity/sicc2021.png',
-  },
-  {
-    id: '3',
-    name: '第一届SICC服务创新大会',
-    url: '/imgs/activity/sicc2019.png',
-  },
-  {
-    id: '4',
-    name: '第三届SICC服务创新大会',
-    url: '/imgs/activity/sicc2021.png',
-  },
-  {
-    id: '5',
-    name: '2019年SICC服务设计创新大会',
-    url: '/imgs/activity/sicc2019.png',
-  },
-  {
-    id: '6',
-    name: '2021年SICC服务设计创新大会',
-    url: '/imgs/activity/sicc2021.png',
-  },
-]
-
-// 票类场次数据
-export const tickets = [
-  { id: '1', activityId: 1, date: '2023年5月15日' },
-  { id: '2', activityId: 1, date: '2023年5月16日' },
-  { id: '3', activityId: 1, date: '2023年5月17日' },
-  { id: '4', activityId: 2, date: '2023年6月20日' },
-  { id: '5', activityId: 2, date: '2023年6月21日' },
-  { id: '6', activityId: 3, date: '2019年11月10日' },
-  { id: '7', activityId: 4, date: '2021年3月16日' },
-  { id: '8', activityId: 4, date: '2021年3月17日' },
-  { id: '9', activityId: 5, date: '2023年7月20日' },
-  { id: '10', activityId: 5, date: '2023年7月21日' },
-  { id: '11', activityId: 6, date: '2023年6月1日' },
-  { id: '12', activityId: 6, date: '2023年6月2日' },
-  { id: '13', activityId: 7, date: '2023年5月12日' },
-  { id: '14', activityId: 7, date: '2023年5月13日' },
-  { id: '15', activityId: 8, date: '2023年11月10日' },
-  { id: '16', activityId: 9, date: '2023年8月16日' },
-  { id: '17', activityId: 10, date: '2023年4月16日' },
-  { id: '18', activityId: 11, date: '2023年4月22日' },
-  { id: '19', activityId: 12, date: '2023年3月16日' },
-]
-
-// 票档价格数据
-export const prices = [
-  { id: '1', activityId: 1, description: '早鸟价-单人票', price: 98, originalPrice: 118, person: 1 },
-  { id: '2', activityId: 1, description: '早鸟价-双人票', price: 178, originalPrice: 218, person: 2 },
-  { id: '3', activityId: 1, description: '正价-单人票', price: 118, originalPrice: 118, person: 1 },
-  { id: '4', activityId: 1, description: '正价-双人票', price: 218, originalPrice: 218, person: 2 },
-
-  { id: '5', activityId: 2, description: '早鸟价-单人票', price: 128, originalPrice: 168, person: 1 },
-  { id: '6', activityId: 2, description: '早鸟价-双人票', price: 228, originalPrice: 288, person: 2 },
-  { id: '7', activityId: 2, description: '正价-单人票', price: 168, originalPrice: 168, person: 1 },
-  { id: '8', activityId: 2, description: '正价-双人票', price: 288, originalPrice: 288, person: 2 },
-
-  { id: '9', activityId: 3, description: '早鸟价-单人票', price: 88, originalPrice: 118, person: 1 },
-  { id: '10', activityId: 3, description: '早鸟价-双人票', price: 168, originalPrice: 218, person: 2 },
-  { id: '11', activityId: 3, description: '正价-单人票', price: 118, originalPrice: 118, person: 1 },
-  { id: '12', activityId: 3, description: '正价-双人票', price: 218, originalPrice: 218, person: 2 },
-
-  { id: '13', activityId: 4, description: '早鸟价-单人票', price: 108, originalPrice: 128, person: 1 },
-  { id: '14', activityId: 4, description: '早鸟价-双人票', price: 208, originalPrice: 228, person: 2 },
-  { id: '15', activityId: 4, description: '正价-单人票', price: 128, originalPrice: 128, person: 1 },
-  { id: '16', activityId: 4, description: '正价-双人票', price: 228, originalPrice: 228, person: 2 },
-
-  { id: '17', activityId: 5, description: '早鸟价-单人票', price: 88, originalPrice: 118, person: 1 },
-  { id: '18', activityId: 5, description: '早鸟价-双人票', price: 168, originalPrice: 218, person: 2 },
-  { id: '19', activityId: 5, description: '正价-单人票', price: 118, originalPrice: 118, person: 1 },
-  { id: '20', activityId: 5, description: '正价-双人票', price: 218, originalPrice: 218, person: 2 },
-
-  { id: '21', activityId: 6, description: '免费活动', price: 0, originalPrice: 0, person: 1 },
-
-  { id: '22', activityId: 7, description: '早鸟价-单人票', price: 78, originalPrice: 128, person: 1 },
-  { id: '23', activityId: 7, description: '早鸟价-双人票', price: 158, originalPrice: 228, person: 2 },
-  { id: '24', activityId: 7, description: '正价-单人票', price: 128, originalPrice: 128, person: 1 },
-  { id: '25', activityId: 7, description: '正价-双人票', price: 228, originalPrice: 228, person: 2 },
-
-  { id: '26', activityId: 8, description: '早鸟价-单人票', price: 88, originalPrice: 118, person: 1 },
-  { id: '27', activityId: 8, description: '正价-单人票', price: 118, originalPrice: 118, person: 1 },
-
-  { id: '28', activityId: 9, description: '早鸟价-单人票', price: 108, originalPrice: 128, person: 1 },
-  { id: '29', activityId: 9, description: '正价-单人票', price: 128, originalPrice: 128, person: 1 },
-
-  { id: '30', activityId: 10, description: '免费活动', price: 0, originalPrice: 0, person: 1 },
-
-  { id: '31', activityId: 11, description: '早鸟价-单人票', price: 90, originalPrice: 100, person: 1 },
-  { id: '32', activityId: 11, description: '正价-单人票', price: 100, originalPrice: 100, person: 1 },
-
-  { id: '33', activityId: 12, description: '早鸟价-单人票', price: 300, originalPrice: 500, person: 1 },
-  { id: '34', activityId: 12, description: '正价-单人票', price: 500, originalPrice: 500, person: 1 },
-]
-
-export const activityList = [
-  {
-    id: 1,
-    title: '2019 SICC服务设计创新大会',
-    cover: '/imgs/activity/cover1.png',
-    score: 5,
-    minPrice: 0,
-    maxPrice: 0,
-    date: new Date(2024, 3, 10),
-    domain: ['IT互联网', '科技'],
-    type: '讲座',
-  },
-  {
-    id: 2,
-    title: '2021 SICC服务设计创新大会',
-    cover: '/imgs/activity/cover2.png',
-    score: 4.5,
-    minPrice: 88,
-    maxPrice: 228,
-    date: new Date(2024, 2, 12),
-    domain: ['IT互联网', '科技'],
-    type: '讲座',
-  },
-
-  {
-    id: 3,
-    title: '少年与星空 插画巡展',
-    cover: '/imgs/activity/cover3.png',
-    score: 4.5,
-    minPrice: 98,
-    maxPrice: 118,
-    date: new Date(2025, 2, 15),
-    domain: ['艺术设计'],
-    type: '展览',
-  },
-  {
-    id: 4,
-    title: 'Universe AI艺术展',
-    cover: '/imgs/activity/cover4.png',
-    score: 3.5,
-    minPrice: 128,
-    maxPrice: 228,
-    date: new Date(2024, 8, 10),
-    domain: ['艺术设计'],
-    type: '展览',
-  },
-  {
-    id: 5,
-    title: '第一届SICC服务创新大会',
-    cover: '/imgs/activity/sicc2019.png',
-    score: 3.5,
-    minPrice: 98,
-    maxPrice: 118,
-    date: new Date(2023, 6, 20),
-    domain: ['IT互联网', '科技'],
-    type: '讲座',
-  },
-  {
-    id: 6,
-    title: '星空与少年 插画巡展',
-    cover: '/imgs/activity/cover3.png',
-    score: 3.5,
-    minPrice: 0,
-    maxPrice: 0,
-    date: new Date(2023, 5, 1),
-    domain: ['艺术设计'],
-    type: '展览',
-  },
-  {
-    id: 7,
-    title: '第三届SICC服务创新大会',
-    cover: '/imgs/activity/sicc2021.png',
-    score: 3.5,
-    minPrice: 88,
-    maxPrice: 228,
-    date: new Date(2023, 4, 12),
-    domain: ['IT互联网', '科技'],
-    type: '讲座',
-  },
-  {
-    id: 8,
-    title: '前端技术深度工作坊',
-    cover: 'https://tdesign.gtimg.com/mobile/demos/swiper1.png',
-    score: 3.5,
-    minPrice: 98,
-    maxPrice: 118,
-    date: new Date(2023, 10, 10),
-    domain: ['IT互联网'],
-    type: '工作坊',
-  },
-  {
-    id: 9,
-    title: 'AI赋能未来设计峰会',
-    cover: 'https://tdesign.gtimg.com/mobile/demos/swiper2.png',
-    score: 3.5,
-    minPrice: 128,
-    maxPrice: 228,
-    date: new Date(2023, 7, 16),
-    domain: ['IT互联网', '科技'],
-    type: '讲座',
-  },
-  {
-    id: 10,
-    title: '数字营销与电商增长大会',
-    cover: 'https://tdesign.gtimg.com/mobile/demos/example1.png',
-    score: 3,
-    minPrice: 0,
-    maxPrice: 0,
-    date: new Date(2023, 3, 16),
-    domain: ['电商'],
-    type: '讲座',
-  },
-  {
-    id: 11,
-    title: '现代摄影艺术展',
-    cover: 'https://tdesign.gtimg.com/mobile/demos/example2.png',
-    score: 3.5,
-    minPrice: 100,
-    maxPrice: 100,
-    date: new Date(2023, 3, 22),
-    domain: ['摄影', '艺术设计'],
-    type: '展览',
-  },
-
-  {
-    id: 12,
-    title: '教育科技创新论坛',
-    cover: 'https://tdesign.gtimg.com/mobile/demos/example3.png',
-    score: 3,
-    minPrice: 350,
-    maxPrice: 500,
-    date: new Date(2023, 2, 16),
-    domain: ['教育', '科技'],
-    type: '讲座',
-  },
-]
