+++ src/mocks/db.ts
@@ -2,27 +2,9 @@ import { fakerZH_CN as faker } from '@faker-js/faker'
 import { factory, manyOf, primaryKey } from '@mswjs/data'
 import { defaultFilterOptions } from '@/constant/filters'
 import { generateDateRange } from '@/utils/dateTime'
-import { covers } from './activityMocks'
+import { cnCommentPool, cnIntroducePool, covers } from './mockData'
 
 const [sixMonthsAgo, sixMonthsLater] = generateDateRange()
-
-// 中文介绍文案池
-const cnIntroducePool = [
-  '在数字化时代背景下，如何抓住机遇，构建“数字+文化”更高效、宽领域、深覆盖的文化创新之路，让优秀传统文化得以延续与新生，是我们持续探索的方向。',
-  '活动将围绕文化保育与文化创新，邀请来自文物、建筑、服饰、工艺等不同艺术领域的专家，分享最新趋势与实践经验。',
-  '通过跨界交流与案例解析，帮助来宾在传统文化与数字科技的碰撞中获得启发，探索服务设计的新命题。',
-  '现场还将设置多种互动环节与展陈内容，呈现真实的服务设计实践，助力参与者获得系统的认知与方法论。',
-]
-
-// 中文评论文案池
-const cnCommentPool = [
-  '活动组织得很专业，内容充实有干货，期待下次再参加。',
-  '嘉宾分享观点独到，案例贴近实际，受益匪浅。',
-  '现场互动环节很有趣，整体节奏把控得很好。',
-  '信息量大且结构清晰，对我的工作有很大启发。',
-  '场地与服务体验不错，细节考虑得很周到。',
-  '主题设置紧贴行业趋势，收获了许多新思路。',
-]
 
 function generatePrice() {
   const priceType = faker.helpers.arrayElement(['free', 'single', 'range'])
