+++ src/pages/main/user.vue
@@ -45,7 +45,7 @@ async function fetchActivities() {
     const pageData = Array.isArray(res.data) ? res.data : []
     activities.value = pageData.map(it => ({
       id: String(it.id),
-      cover: it.cover ?? '/mock-imgs/activity/cover1.png',
+      cover: it.cover,
       title: it.title,
       time: formatDate(it.date),
       status: it.status,
